+++ src/render/pipeline/buffer/mod.rs
@@ -20,14 +20,18 @@ impl From<BufferKind> for wgpu::BufferUsages {
 
 // A custom buffer container for dynamic resizing.
 pub struct Buffer {
-    pub raw: wgpu::Buffer,
+    raw: wgpu::Buffer,
     label: &'static str,
     usage: wgpu::BufferUsages,
     size_of_type: u64,
-    pub count: u32,
+    count: usize,
 }
 
 impl Buffer {
+    pub fn len(&self) -> usize {
+        self.count
+    }
+
     pub fn new<T>(device: &wgpu::Device, label: &'static str, kind: BufferKind) -> Self {
         let size_of_type = std::mem::size_of::<T>() as u64;
         let usage = kind.into();
@@ -45,7 +49,7 @@ impl Buffer {
         }
     }
 
-    pub fn resize(&mut self, device: &wgpu::Device, new_count: u32) {
+    pub fn resize(&mut self, device: &wgpu::Device, new_count: usize) {
         self.raw = device.create_buffer(&wgpu::BufferDescriptor {
             label: Some(self.label),
             size: self.size_of_type * new_count as u64,
@@ -57,5 +61,17 @@ impl Buffer {
 
     pub fn write_data<T: bytemuck::Pod>(&mut self, queue: &wgpu::Queue, data: &T) {
         queue.write_buffer(&self.raw, 0, bytemuck::bytes_of(data));
+    }
+
+    pub fn write_slice<A: bytemuck::NoUninit>(&mut self, queue: &wgpu::Queue, slice: &[A]) {
+        queue.write_buffer(&self.raw, 0, bytemuck::cast_slice(slice));
+    }
+
+    pub fn raw_slice(&self) -> wgpu::BufferSlice<'_> {
+        self.raw.slice(..)
+    }
+
+    pub fn binding_resource(&self) -> wgpu::BindingResource<'_> {
+        self.raw.as_entire_binding()
     }
 }
